refactor(ai-builder): Extract eval harness runner domain modules behind a re-export facade (no-changelog)
Move every runner.ts export verbatim into domain modules (build-workflow,
scenario-execution, agent-execution, seed-tables, cleanup); runner.ts
becomes a pure re-export facade so importers stay stable until repointed.
